% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/dp_generators.R
\name{dp_generator}
\alias{dp_generator}
\title{A R function that returns a generator for the dirichlet process}
\usage{
dp_generator(alpha, base)
}
\arguments{
\item{alpha}{The concentration parameter for the dirichlet process}

\item{base}{The base distribution (commonly refered to as H) for the dirichlet process}
}
\description{
Returns a generator for a dirichlet process (DP) of a specified alpha and base distribution.
}
\examples{
base <- function(n) rnorm(n, mean=5, sd=2)
dp <- dp_generate(2, base)
dp(5)
}
\keyword{dirichlet}
\keyword{process}

