% Generated by roxygen2 (4.0.2): do not edit by hand
\name{infinite_cluster}
\alias{infinite_cluster}
\title{Infinite dimensional clustering for R}
\usage{
infinite_cluster(data, alpha, distance = "Euclidian")
}
\arguments{
\item{data}{The data to be clustered by the algorithm}

\item{alpha}{The dispersion parameter for the clustering algorithm}

\item{distance}{The distance metric used in the clustering algorithm. The default is set to Euclidean distance,
but other options include: manhattan, mahalanobis, and cosine distance}
}
\description{
Performs infinite dimensional clustering, as specified by Algorithm 1 in
 "___" by Jordan, Blei, et al. Unlike the typical k-means
clustering algorithm when the value of *k* is specified, this algorithm clusters the observations according to a dispersion
parameter (alpha).
}
\examples{
output <- infinite_cluster(data, alpha=2, distance="Euclidian")
}
\keyword{clustering,}
\keyword{dp}
\keyword{infinite,}

