% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mondrian_2d}
\alias{mondrian_2d}
\title{Multi-dimensional Mondrian Process}
\usage{
mondrian_2d(budget, domain = matrix(c(0, 0, 1, 1), nrow = 2))
}
\arguments{
\item{budget}{The budget for the Mondrian Process}

\item{domain}{A matrix/dataframe of intervals for the mondrian process. The default is set to [0,1][0,1]}
}
\description{
Returns the split locations for all dimensions in a multi dimensional mondrian process,
from a specified domain matrix and a budget. More information about this can be found in
(http://danroy.org/papers/RoyTeh-NIPS-2009.pdf).
}
\examples{
splits <- mondrian_2d(5)
}
\keyword{mondrian}
\keyword{process}

