% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/dp_mm.R
\name{plot.dp_mm}
\alias{plot.dp_mm}
\title{Plot the Dirichlet Process Mixture Model output returned from the 'dp_mm' function}
\usage{
\method{plot}{dp_mm}(obj, ...)
}
\arguments{
\item{obj}{The output of the 'dp_mm' function}
}
\description{
Returns a ggplot of the dirichlet process mixture model, with each of the clusters colored accordingly
}
\examples{
obj <- dp_mm(x, n_sweeps=100)
plot(obj, main = "My plot")
}
\keyword{dpmm}
\keyword{plot,}

