% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/infinite_cluster.R
\name{plot.infinite_mm}
\alias{plot.infinite_mm}
\title{Plot the output of the infinite mixture model}
\usage{
\method{plot}{infinite_mm}(output)
}
\arguments{
\item{output}{The S3 object returned from the infinite_cluster function}
}
\description{
Plot Plot PLot
}
\keyword{plot}

