% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_mondrian}
\alias{plot_mondrian}
\title{Plots of Mondrian Process, for visualization purposes}
\usage{
plot_mondrian(budget, interval = c(0, 1), pretty = TRUE...)
}
\arguments{
\item{budget}{The budget for the mondrian process}

\item{interval}{The interval for the mondrian process}

\item{pretty}{An option to add color and text to the visualization}
}
\description{
A R function that plots of the Mondrian process for a specified budget value and interval ranges.
Currently only supports the one dimensional and two dimensional cases
}
\examples{
plot_mondrian(5) # single dimension
plot_mondrian(5, interval=matrix(c(0,0,1,1), nrow=2)) # two dimensions
}
\keyword{mondrian}
\keyword{plot,}

