% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/infinite_cluster.R
\name{predict.infinite_mm}
\alias{predict.infinite_mm}
\title{Classify a new observation with the infinite mm}
\usage{
\method{predict}{infinite_mm}(new_obs, model)
}
\arguments{
\item{new_obs}{A vector of observations to classify}

\item{model}{The output from the infinite_mm function (S3 object)}
}
\description{
Picks the closest cluster for a new observation, or a set of observations
}
\keyword{infinite_mm}
\keyword{predict,}

