% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/dp_generators.R
\name{py_generator}
\alias{py_generator}
\title{A generator for the Pitman-Yor Process in R}
\usage{
py_generator(alpha, d, base)
}
\arguments{
\item{alpha}{The concentration parameter}

\item{d}{The tuning paramter for the Pitman-Yor. If d=1 we have the Dirichlet Process}

\item{base}{The base distribution for the Pitman-Yor process}
}
\description{
Returns a generator for a Pitman-Yor process, given a specified $d$, $alpha$, and a base distribution
}
\examples{
base <- function(n) rnorm(n, mean=5, sd=2)
py <- py_generate(3, 2, base)
py(5)
}
\keyword{pitman-yor,}
\keyword{process}

