% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/dp_generators.R
\name{py_sample}
\alias{py_sample}
\title{Sample for the Pitman-Yor Process}
\usage{
py_sample(n, alpha, d, base)
}
\arguments{
\item{n}{The number of samples to sample}

\item{alpha}{The concentration parameter for the PY process}

\item{d}{The tuning parameter for the PY process}

\item{base}{The base distribution (commonly refered to as H) for the PY process}
}
\description{
Returns a sample (non-unique) from the pitman-yor process with a specified
concentration parameter (alpha), tuning paramter (d) and a base distribution
}
\examples{
base <- function(n) rpois(n, 20)
samp <- py_sample(10, 5, 2, base)
}
\keyword{pitman-yor}
\keyword{process}

