% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/dp_mm.R
\name{summary.dp_mm}
\alias{summary.dp_mm}
\title{Get a summary for the dirichlet process mixture model}
\usage{
\method{summary}{dp_mm}(obj)
}
\arguments{
\item{obj}{The output of the 'dp_mm' function.}
}
\description{
This function returns a summary of the dirichlet process mixture model. The summary includes the cluster parameters
at each iteration, the cluster labels, and more information about the mixture model.
}
\examples{
obj <- dp_mm(x, n_sweeps=100)
summary(obj)
}
\keyword{dpmm}
\keyword{summary,}

