% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/dp_mm.R
\name{dp_mm}
\alias{dp_mm}
\title{Dirichlet Process Mixture Modeling with Gibbs Sampling}
\usage{
dp_mm(x, alpha)
}
\arguments{
\item{x}{The data (a data.frame) to be passed in to the DP mixture model}

\item{n_sweeps}{The number of iterations to perform for the Gibbs Sampler}
}
\description{
Implements a Dirichlet Process Mixture Model with Gibbs Sampling. The base distribution for the Dirichlet
Process is a Normal-Inverse-Gamma distribution, which results in a predictive posterior of the t-distribution. After
Gibbs Sampling, this function returns a cluster parameters for each of the clusters, as the Dirichlet Process Mixture Model
is a non-parametric clustering algorithm. The alpha (concentration parameter for the DP) is resampled after each iteration
using the soft-max trick, and each of the cluster parameters are resampled at each iteration as well.
}
\examples{
output <- dp_mm(x, n_sweeps = 100)
summary(output)
plot(output)
}
\keyword{clustering}
\keyword{dirichlet}
\keyword{mixture}
\keyword{model,}
\keyword{process,}

