% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.dp_mm}
\alias{plot.dp_mm}
\title{Plot the Dirichlet Process Mixture Model output returned from the 'dp_mm' function}
\usage{
\method{plot}{dp_mm}(obj, ...)
}
\arguments{
\item{obj}{The output of the 'dp_mm' function}
}
\description{
Returns a ggplot of the dirichlet process mixture model, with each of the clusters colored accordingly
}
\examples{
# Make up some data
data <- c(rnorm(50, 4, 1), rnorm(50, 8, .5))
output <- dp_mm(data, 3, num_iters=100)
plot(obj, col="red", lty=3)
}
\keyword{dpmm}
\keyword{plot,}

