% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict.infinite_mm}
\alias{predict.infinite_mm}
\title{Classify a new observation with the infinite mm}
\usage{
\method{predict}{infinite_mm}(new_obs, model)
}
\arguments{
\item{new_obs}{A vector of observations to classify}

\item{model}{The output from the infinite_mm function (S3 object)}
}
\description{
Picks the closest cluster for a new observation, or a set of observations
}
\keyword{infinite_mm}
\keyword{predict,}

