% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.dp_mm}
\alias{summary.dp_mm}
\title{Get a summary for the dirichlet process mixture model}
\usage{
\method{summary}{dp_mm}(obj)
}
\arguments{
\item{obj}{The output of the 'dp_mm' function.}
}
\description{
This function returns a summary of the dirichlet process mixture model. The summary includes the cluster parameters
at each iteration, the cluster labels, and more information about the mixture model.
}
\examples{
# Make up some data
data <- c(rnorm(50, 4, 1), rnorm(50, 8, .5))
output <- dp_mm(data, 3, num_iters=100)
summary(output)
}
\keyword{dpmm}
\keyword{summary,}

